Avoid eager torch imports in CLI config paths
