+++ vaannotate/vaannotate_ai_backend/config.py
@@ -7,7 +7,7 @@ from dataclasses import dataclass, field
 from pathlib import Path
 from typing import List, Optional, Tuple
 
-from .core.embeddings import IndexConfig
+from .core.index_config import IndexConfig
 
 
 def _env_int(name: str, default: Optional[int] = None) -> Optional[int]:
